separated plans page into plan card and page
